Fix bug with allocation of memory.

+++ src/stack.h
@@ -5,7 +5,7 @@
 
 #define VAR_INFO(var) ((var_info_t) {__FILE__, __LINE__, __PRETTY_FUNCTION__, #var})
 
-typedef unsigned long long elem_t;
+typedef int elem_t;
 
 const size_t DEF_STACK_CAPACITY = 10;
 const elem_t DATA_POISON = (elem_t) 0xDEADDA1A;
